support for hidden google maps
googlemaps can not initialize if it is not visible ... since entwine does not support
the :has selector and we do not know in which tab the map is located, the only way to
get this working (afaik) is to re-init the current tab on every tab switch. To limit the
performance loss the init function will only act on googlemap fields in non hidden tabs.